// apps/angular-app/src/app/rest-owner/shared/models/order-detail.ts
import { CouponDetail } from "./coupon-detail";

export class OrderDetail {
    ID: number;
    OrderNumber: string;
    SubTotal: number;
    DeliveryCharge: number;
    ServiceFee: number;
    RestaurantWaivedFee: number;
    Total: number;
    Driver_Tip_Percent: number;
    Driver_Tip: number;
    Owner_ServiceFee_Rate: number;
    Owner_DeliveryCharge_Restaurant: number;
    Owner_NetAccountCredit: number;
    ServiceFee_TransactionFee: number;
    GrandTotal: number;
    Apartment_Suite: string;
    Instructions: string;
    Tax_Rate: number;
    Tax: number;
    Is_Accepted: boolean;
    Is_Cancelled_By_Customer: boolean;
    Is_Cancelled_By_Restaurant: boolean;
    PaymentMethodID: number;
    PaymentMethod: string;
    OrderDate: string;
    CustomerName: string;
    ServiceType: string;
    DeliveryModeID: number;
    DeliveryMode: string;
    DeliveryAddress: string;
    RestaurantAddress: string;
    RestPhone: string;
    RestEmail: string;
    CustomerID: number;
    ServiceTypeID: number;
    DeliveryAddressID: number;
    RestaurantID: number;
    Restaurant: string;
    AgentID: any;
    StatusID: number;
    Status: string;
    OrderTimeDetails: string;
    OrderETA: string;
    Owner_SubscriptionID: number;
    SubscriptionName: string;
    SoldOutAction: string;
    Is_Read: boolean;
    CouponDetail: Array<CouponDetail>;
}